Doc: Fractional knapsack
Related Work Items: #779, #797

+++ FractionalKnapsack.cs
@@ -5,8 +5,13 @@ namespace Greedy
 {
     public class FractionalKnapsack
     {
+        /// <summary>Calculates maximum value that can be filled in knapsack.</summary>
+        /// <param name="items">The items.</param>
+        /// <param name="knapCapacity">Knapsack capacity.</param>
+        /// <returns>Maximum value that can be filled in knapsack.</returns>
         private int CalculateMaxValue(Item[] items, int knapCapacity)
         {
+            // Sort based on cost/kg in descending order
             Array.Sort(items, (i1, i2) => -1 * i1.ValuePerKg.CompareTo(i2.ValuePerKg));
 
 
@@ -23,13 +28,13 @@ namespace Greedy
                 {
                     WriteLine(items[i]);   // Print full item
 
-                    knapCap -= items[i].Weight;
-                    return items[i].Value + FindMaxItem(i + 1, knapCap);
+                    knapCap -= items[i].Weight;   // Remaining capacity
+                    return items[i].Value + FindMaxItem(i + 1, knapCap);   // Add next item
                 }
 
                 // Find partial item quantity & value
-                int weight = knapCap,
-                    value = (int) Math.Round(items[i].ValuePerKg * weight);
+                int weight = knapCap,   // Remaining capacity = item's partial weight
+                    value = (int)Math.Round(items[i].ValuePerKg * weight);   // Partial weight value
                 WriteLine($"{items[i].Name}: {weight} Kg, {value:C0}");   // Print partial item
                 return value;
             }
@@ -44,9 +49,9 @@ namespace Greedy
                 new Item("Item 1", 6, 6),
                 new Item("Item 2", 10, 2),
                 new Item("Item 3", 3, 1),
-                new Item("Item 4", 5,8 ),
-                new Item("Item 5", 1,3 ),
-                new Item("Item 6", 3,5 )
+                new Item("Item 4", 5, 8),
+                new Item("Item 5", 1, 3),
+                new Item("Item 6", 3, 5)
 
                 //new Item("Item 1", 20, 100),
                 //new Item("Item 2", 30, 120),
@@ -62,14 +67,19 @@ namespace Greedy
 
 
 
+    /// <summary>Struct Item</summary>
     internal struct Item
     {
+        /// <summary>Initializes a new instance of the <see cref="T:Greedy.Item"/> struct.</summary>
+        /// <param name="name">The name.</param>
+        /// <param name="weight">The weight.</param>
+        /// <param name="value">The value.</param>
         internal Item(string name, int weight, int value)
         {
             Name = name;
             Weight = weight;
             Value = value;
-            ValuePerKg = (float) Value / (float) Weight;
+            ValuePerKg = (float)Value / Weight;   // cost/kg
         }
 
         public string Name { get; }
